Day 17: Enhance meshlet and triangle culling
We now implement frustum culling of individual meshlets and
backface+small primitive culling for individual triangles (in mesh
shader).

The former is unambiguously good, the latter may or may not be
worthwhile depending on a complex set of hardware conditions &
meshlet/mesh shader setup, which we do *not* do optimally at this time.