Day 20: Implement and switch to task command submission
This mode dispatches a single draw call instead of using MDI,
which dramatically boosts efficiency when we are not bottlenecked
by mesh shading or rasterization work. This allows us to save
A LOT of time by cluster occlusion culling, which we had to
keep off by default because of these issues.

Since new mode seems to always be as good or better than old mode,
switch to it unconditionally and clean up some code. We also
enable cluster occlusion by default now that it's actually beneficial.