Day 22: Initial glTF scene support and Implement scene camera support
We now can load mesh data and node data from glTF files with some
restrictions, and render that instead of instanced .obj files!
glm had to be reconfigured to fix constructor argument order.

We now use the camera present in glTF file to render the scene; if
absent, we still use the camera at origin, but flip it to point to -Z
which was required to align the behavior with glTF.