Refactor: Minor shading tweaks, Switch to float16 positions and Move vertexOffset from MeshDraw into Meshlet
- Add camera position and basic specular with fixed power
- Use cull mask 0xff for now as it's faster on amdvlk
- In theory float16 may not be enough for some large meshes, but for now
just switch to it unconditionally; do this in a separate commit so that
we can revert it later if need be.

This reduces the vertex size from 24 bytes to 20 bytes; we can
separately go down to 16 bytes by packing TBN better.
- This still keeps Meshlet size the same due to alignment and is just as
fast, but makes the code easier to modify as MeshDraw no longer is tied
to how geometry is laid out.