Day 27: Optimize shadow traversal a little by casting two rays
This relies on efficient object filtering based on cull opaque flags in
RT implementations; for example, currently on radv two traces are much
slower than one as it performs this filtering too late.

Also fix geometry flags to be unspecified; this should not matter as
long as TLAS overrides them, but is just cleaner.