Add basic tonemapping, Move shade.comp.glsl to final.comp.glsl, Disable temporal checkerboarding for now
After some experimentation for now go with the tried-and-true filmic
option. Needed to tweak the lighting a little bit for that, and also fix
the specular that was not taking sun intensity into account.

This matches the internal names better and is easier to navigate to.

This is not really usable without a proper temporal filter with reuse.