Refactor: Remove deprecated android main and remove use of SPIRV shader modules on Windows
Vulkan 1.4 (via maintenance6) allows us to stop using SPIRV shader
module objects and pass the SPIRV data inline when creating pipelines.

This is particularly useful because we don't need to filter out SPIRV
modules with features that the GPU doesn't support (like mesh shaders or
raytracing); the module blobs will only be parsed when creating the
pipelines and these won't be created if the features are not supported.